Explicitly re-raise exceptions in handlers
Starting in OCaml 4.02, the compiler detects whether an exception is
being re-raised by looking for the exact text `raise e` (when the
exception was caught using `with e ->`. This is fragile and doesn't
work at all with Async Monitors, so just explicitly `reraise`
everywhere that we do it.

This should dramatically improve backtraces thrown inside of
transactions.