Make Pgx_value.t opaque
It's not safe to use the string contents directly, so this makes
Pgx_value.t opaque.

The biggest issue here is that the string and binary representations
can be different for the same data, but there are weird issues
with using the raw representation for floats and other conversions
as well.