Add transaction sub-sequencer
This ensures that if you start a transaction, other queries outside
of it have to wait for the transaction to finish.

For example, if you do:

```
let%bind _ = Pgx_async.with_transaction (...)
and _ = Pgx_async.execute ...
```

This will now guarantee that the execute happens outside of the
transaction (either before or after it, but not during).